+++ src/app/(app)/(main)/pricing/layout.tsx
@@ -1,8 +1,7 @@
 import type { Metadata } from 'next';
 
 export const metadata: Metadata = {
-  title: 'Url Expired',
-  description: 'The url you are trying to access has expired',
+  title: 'Pricing | UrlHub',
 };
 
 export default function Layout({